refactor: replace controllers with routers
