+++ apps/email/src/container.ts
@@ -2,6 +2,7 @@ import { ConfigService } from 'src/services/config.service.js';
 import { LoggerService } from '@weather-subscription/shared';
 import { TemplateService } from './services/template.service.js';
 import { EmailService } from './services/email.service.js';
+import { EmailRouter } from './routers/email.router.js';
 import { Server } from './server.js';
 import { App } from './app.js';
 
@@ -13,6 +14,8 @@ export class Container {
 
   public templateService = new TemplateService();
   public emailService = new EmailService(this);
+
+  public emailRouter = new EmailRouter(this);
 
   public server = new Server(this);
