+++ apps/gateway/src/routers/ui.router.tsx
@@ -1,0 +1,12 @@
+import { createRoute, OpenAPIHono } from '@hono/zod-openapi';
+import { RootView } from '../views/root.view.js';
+
+export class UiRouter {
+  public setup(app: OpenAPIHono) {
+    // if you specify the 200 schema as a string, you cannot be able to use c.html because of type issues
+    app.openapi(
+      createRoute({ method: 'get', path: '/', responses: { 200: { description: 'the root page' } } }),
+      async (ctx) => await ctx.html(<RootView />),
+    );
+  }
+};
